examples/3-deepresearch/agent-code/pyproject.toml
[project]
name = "deepresearch"
version = "0.1.0"
description = "Deep research agent powered by Strands DeepAgents with web search capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.0.7",
    "linkup-sdk>=0.9.0",
    "strands-agents>=1.18.0",
    "strands-agents-tools>=0.2.16",
    "strands-deep-agents>=0.1.1",
]

[project.optional-dependencies]
# Optional telemetry with Weave
weave = [
    "weave>=0.50.0",
]
# OpenTelemetry support
otel = [
    "strands-agents[otel]>=1.18.0",
]
# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "moto[s3,secretsmanager]>=5.0.0",
    "bedrock-agentcore-starter-toolkit>=0.1.34",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["deepresearch"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
